refactor(mesh): rename incremental.rs to collector.rs
The file no longer holds anything "incremental" — IncrementalUpdateCollector
was deleted, leaving CentralCollector, PeerWatermark, and RoundBatch.
Rename the module to match what it actually contains.

Updates:
- git mv incremental.rs collector.rs
- lib.rs: mod incremental -> mod collector
- controller.rs / ping_server.rs / sync.rs: use crate::collector::...

Variables and log messages mentioning "incremental update" are kept —
they describe the wire-level behavior (sending deltas, not full dumps),
which is still accurate.
